# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EasyKos API Documentation",
        "description": "Dokumentasi API untuk aplikasi EasyKos - Manajemen Kos-kosan",
        "contact": {
            "email": "admin@easykos.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "EasyKos API Server Local"
        }
    ],
    "paths": {
...
        "/api/ulasan": {
            "get": {
                "tags": [
                    "Ulasan"
                ],
                "summary": "Daftar ulasan (publik)",
                "operationId": "8f24bcc776fb5d763b35f77b3da136a6",
                "parameters": [
                    {
                        "name": "kosan_id",
                        "in": "query",
                        "description": "Filter ulasan berdasarkan ID Kosan",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar ulasan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ulasan"
                ],
                "summary": "Tambah ulasan untuk kosan (harus pernah menyewa)",
                "operationId": "bce09c0f56cb8dc3bd97752c0ebedf56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kosan_id",
                                    "rating"
                                ],
                                "properties": {
                                    "kosan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "komentar": {
                                        "type": "string",
                                        "example": "Tempatnya bersih dan nyaman."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ulasan berhasil ditambahkan"
                    },
                    "403": {
                        "description": "Hanya penyewa yang bisa mengulas"
                    },
                    "409": {
                        "description": "Sudah pernah mengulas"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/ulasan/{id}": {
            "get": {
                "tags": [
                    "Ulasan"
                ],
                "summary": "Detail ulasan (publik)",
                "operationId": "5a447212ba78938e38500aa615a6de2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Ulasan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail ulasan"
                    },
                    "404": {
                        "description": "Ulasan tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ulasan"
                ],
                "summary": "Hapus ulasan milik sendiri atau admin",
                "operationId": "9b3ee29a919329c9b2a0d9ae0c73326f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Ulasan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ulasan berhasil dihapus"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Ulasan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Ulasan"
                ],
                "summary": "Update ulasan milik sendiri",
                "operationId": "7f0183052947600a2c996f1cc06b88e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Ulasan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "komentar": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ulasan berhasil diperbarui"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Ulasan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT Anda di sini. Cukup masukkan tokennya saja (tanpa kata 'Bearer ')",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}